calendarioproduccion/backend/requirements.txt
Flask==3.0.3
Flask-Cors==5.0.0
SQLAlchemy==2.0.32
python-dotenv==1.0.1
